backend/requirements.txt
Flask
requests
pandas
flask-cors
python-dotenv
PyJWT
